+++ integrations/vespa/src/haystack_integrations/document_stores/vespa/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from .document_store import DEFAULT_BM25_RANKING, DEFAULT_SEMANTIC_RANKING, VespaDocumentStore
+
+__all__ = ["DEFAULT_BM25_RANKING", "DEFAULT_SEMANTIC_RANKING", "VespaDocumentStore"]