+++ integrations/langfuse/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [integrations/langfuse-v5.0.1] - 2026-05-18
+
+### 🐛 Bug Fixes
+
+- Fix langfuse tracer when input messages is `None` (#3324)
+
+
 ## [integrations/langfuse-v5.0.0] - 2026-04-17
 
 ### 🚀 Features